Removing ase dependency on how the structure string is written.
Moving the parsing functions to the actual parser folder to make the structure cleaner.
