+++ aiida_lammps/calculations/base.py
@@ -12,10 +12,10 @@ from aiida import orm
 from aiida.common import datastructures, exceptions
 from aiida.engine import CalcJob
 
-from aiida_lammps.common.generate_structure import generate_lammps_structure
-from aiida_lammps.common.input_generator import generate_input_file
 from aiida_lammps.data.potential import LammpsPotentialData
 from aiida_lammps.data.trajectory import LammpsTrajectory
+from aiida_lammps.parsers.inputfile import generate_input_file
+from aiida_lammps.parsers.utils import generate_lammps_structure
 
 
 class LammpsBaseCalculation(CalcJob):
