+++ aiida_lammps/parsers/parse_raw/__init__.py
@@ -1,0 +1,13 @@
+"""Set of functions to parse the unformatted raw files generated by lammps"""
+
+from .final_data import parse_final_data
+from .lammps_log import parse_logfile
+from .trajectory import create_structure, iter_trajectories, parse_step
+
+__all__ = (
+    "parse_logfile",
+    "parse_final_data",
+    "parse_step",
+    "create_structure",
+    "iter_trajectories",
+)
