+++ aiida_lammps/parsers/parse_raw/lammps_log.py
@@ -1,0 +1,93 @@
+"""Set of functions to parse the unformatted log files generated by LAMMPS.
+"""
+# pylint: disable=fixme
+import ast
+import re
+from typing import Union
+
+import numpy as np
+
+
+def parse_logfile(filename: str = None, file_contents: str = None) -> Union[dict, dict]:
+    """
+    Parse the log.lammps file.
+
+    This will gather the time dependent data stored in the log file and
+    stores it as a dictionary. It will also gather single quantities and stores
+    them into a different dictionary.
+
+    :param filename: name of the lammps log file, defaults to None
+    :type filename: str, optional
+    :param file_contents: contents of the lammps log file, defaults to None
+    :type file_contents: str, optional
+    :return: dictionary with the time dependent data, dictionary with the global data
+    :rtype: Union[dict, dict]
+    """
+    # pylint: disable=too-many-branches, too-many-locals
+
+    if filename is None and file_contents is None:
+        return None
+
+    if filename is not None:
+
+        try:
+            with open(filename) as handler:
+                data = handler.read()
+                data = data.split("\n")
+        except OSError:
+            return None
+
+    if file_contents is not None:
+        data = file_contents.split("\n")
+
+    header_line_position = -1
+    header_line = ""
+    _data = []
+    end_found = False
+    parsed_data = {}
+    global_parsed_data = {}
+
+    perf_regex = re.compile(r".*Performance\:.*\,\s+([0-9\.]*)\stimesteps\/s.*")
+    performance_match = perf_regex.search(file_contents or "")
+    if performance_match:
+        global_parsed_data["steps_per_second"] = float(performance_match.group(1))
+
+    for index, line in enumerate(data):
+        line = line.strip()
+        if "binsize" in line:
+            global_parsed_data["binsize"] = ast.literal_eval(
+                line.split()[2].replace(",", "")
+            )
+            global_parsed_data["bins"] = [
+                ast.literal_eval(entry) for entry in line.split()[5:]
+            ]
+        if "ghost atom cutoff" in line:
+            global_parsed_data["ghost_atom_cutoff"] = ast.literal_eval(line.split()[-1])
+        if "master list distance cutoff" in line:
+            global_parsed_data["master_list_distance_cutoff"] = ast.literal_eval(
+                line.split()[-1]
+            )
+        if "max neighbors/atom" in line:
+            global_parsed_data["max_neighbors_atom"] = ast.literal_eval(
+                line.split()[2].replace(",", "")
+            )
+        if "units" in line:
+            global_parsed_data["units_style"] = line.split()[1]
+        if "Total wall time:" in line:
+            global_parsed_data["total_wall_time"] = line.split()[-1]
+        if "bin:" in line:
+            global_parsed_data["bin"] = line.split()[-1]
+        if line.startswith("Step"):
+            header_line_position = index
+            header_line = [
+                re.sub("[^a-zA-Z0-9_]", "__", entry) for entry in line.split()
+            ]
+        if header_line_position > 0 and index != header_line_position and not end_found:
+            if not line.split()[0].replace(".", "", 1).isdigit():
+                end_found = True
+        if header_line_position > 0 and index != header_line_position and not end_found:
+            _data.append([ast.literal_eval(entry) for entry in line.split()])
+    _data = np.asarray(_data)
+    for index, entry in enumerate(header_line):
+        parsed_data[entry] = _data[:, index].tolist()
+    return {"time_dependent": parsed_data, "global": global_parsed_data}
