+++ aiida_lammps/parsers/utils.py
@@ -1,19 +1,11 @@
-"""
-Creation of the structure file content.
-
-As allowing the users to create their lattices using LAMMPS' would be too
-complex, one must ensure that the aiida StructureData is written to file in
-a format that is compatible to LAMMPS.
-
-In the case of non-orthogonal structures, this will take care of generating
-a triclinic cell compatible with what LAMMPS expects.
-"""
-import typing
+"""Utility functions for the handling of the input files"""
+from collections.abc import Iterable
+from typing import Union
 
 import numpy as np
 
 
-def transform_cell(cell) -> typing.Union[np.array, np.array]:
+def _transform_cell(cell) -> Union[np.array, np.array]:
     """Transform the cell to an orientation, compatible with LAMMPS
 
     LAMMPS requires the simulation cell to be in the format of a
@@ -23,7 +15,7 @@ def transform_cell(cell) -> typing.Union[np.array, np.array]:
 
     :param cell: crystal cell of the original structure
     :returns: LAMMPS compatible cell, transformation between original and final cell
-    :rtype: typing.Union[np.array, np.array]
+    :rtype: Union[np.array, np.array]
     """
     cell = np.array(cell)
     transform, upper_tri = np.linalg.qr(cell.T, mode="complete")
@@ -47,8 +39,15 @@ def generate_lammps_structure(
     charge_dict: dict = None,
     round_dp: float = None,
     docstring: str = "generated by aiida_lammps",
-) -> typing.Union[str, np.array]:
-    """Create lammps input structure file content.
+) -> Union[str, np.array]:
+    """Creation of the structure file content.
+
+    As allowing the users to create their lattices using LAMMPS' would be too
+    complex, one must ensure that the aiida StructureData is written to file in
+    a format that is compatible to LAMMPS.
+
+    In the case of non-orthogonal structures, this will take care of generating
+    a triclinic cell compatible with what LAMMPS expects.
 
     :param structure: the structure to use in the simulation
     :type structure: orm.StructureData
@@ -63,7 +62,7 @@ def generate_lammps_structure(
     :raises ValueError: if the atom_style does not belong to either 'atomic' or 'charge'
     :return: the structure file content, the transformation matrix applied to
         the structure cell and coordinates
-    :rtype: typing.Union[str, np.array]
+    :rtype: Union[str, np.array]
     """
     # pylint: disable=too-many-locals
 
@@ -87,9 +86,12 @@ def generate_lammps_structure(
     filestring += f"{len(structure.sites)} atoms\n"
     filestring += f"{len(kind_name_id_map)} atom types\n\n"
 
-    atoms = structure.get_ase()
-    cell, coord_transform = transform_cell(atoms.cell)
-    positions = np.transpose(np.dot(coord_transform, np.transpose(atoms.positions)))
+    cell, coord_transform = _transform_cell(structure.cell)
+    positions = np.transpose(
+        np.dot(
+            coord_transform, np.transpose([_site.position for _site in structure.sites])
+        )
+    )
 
     if round_dp:
         cell = np.round(cell, round_dp) + 0.0
@@ -124,3 +126,60 @@ def generate_lammps_structure(
             filestring += f" {pos[0]:20.10f} {pos[1]:20.10f} {pos[2]:20.10f}\n"
 
     return filestring, coord_transform
+
+
+def generate_header(value: str) -> str:
+    """
+    Generate the header for the blocks.
+
+    :param value: string indicating the input block
+    :type value: str
+    :return: header/footer for the input block
+    :rtype: str
+    """
+    return "#" + value.center(80, "-") + "#\n"
+
+
+def flatten(full_list: list) -> list:
+    """Flattens a list of list into a flat list.
+
+    :param full_list: list of lists to be flattened
+    :type full_list: list
+    :yield: flattened list
+    :rtype: list
+    """
+    for element in full_list:
+        if isinstance(element, Iterable) and not isinstance(element, (str, bytes)):
+            yield from flatten(element)
+        else:
+            yield element
+
+
+def convert_to_str(value):
+    """convert True/False to yes/no and all values to strings"""
+    if isinstance(value, bool):
+        if value:
+            return "yes"
+        return "no"
+    return str(value)
+
+
+def _convert_values(value):
+    if isinstance(value, (tuple, list)):
+        return " ".join([convert_to_str(v) for v in value])
+    return convert_to_str(value)
+
+
+def join_keywords(dct, ignore=None):
+    """join a dict of {keyword: value, ...} into a string 'keyword value ...'
+
+    value can be a single value or a list/tuple of values
+    """
+    ignore = [] if not ignore else ignore
+    return " ".join(
+        [
+            f"{k} {_convert_values(dct[k])}"
+            for k in sorted(dct.keys())
+            if k not in ignore
+        ]
+    )
