+++ aiida_lammps/parsers/variables_types.json
