+++ tests/test_parsers.py
@@ -8,7 +8,7 @@ from textwrap import dedent
 from aiida.orm import FolderData, SinglefileData
 from aiida.plugins import ParserFactory
 
-from aiida_lammps.common.raw_parsers import parse_final_data, parse_logfile
+from aiida_lammps.parsers.parse_raw import parse_final_data, parse_logfile
 from .utils import TEST_DIR
